simplifier: Evaluate attribute error for both edges for seam collapses
When collapsing seam->seam or seam->locked edges, we were assuming that
the attribute error on both the primary edge and the seam pair is about
the same. There are cases when this is not true; that leads to the
collapse with an ostensibly low error significantly changing attribute
appearance.

To evaluate the error, we need to use the same logic as we use during
perform() to discover the seam pair, and use the maximum error of the
two for the attribute component.