simplifier: Aggregate attribute error for both edges for seam collapses
To get consistent results for seam collapses, we need to accumulate the
error instead of using max/avg. The intuition is that if the seam did
not exist, we would automatically get the aggregation of errors because
we aggregate attribute quadrics; this would result in the error being
scaled with the associated triangle area. When the vertices are split
along the seam edge, we don't get this aggregation for free and must
perform it manually.