simplifier: Slightly unify seam pair discovery
Since we have to unfortunately do seam pair discovery in two places
again, copy the comments and a memory safety guard so that these two
copies align better. performEdgeCollapses still does more validation in
debug here.

As in performEdgeCollapses, the memory safety guard is not actually hit
in practice, but it's not explicitly guarded against anywhere and
instead isn't hit through a complex reasoning chain due to how seams and
loops are structured. Not ideal, but probably still worth keeping for
now.