simplifier: Only evaluate unidirectional changes once
While in many cases most of the edges are bidirectional, some models
have a high number of unidirectional edges, and that number only grew
with unlocked seam/border->locked collapses. We've been using a
branchless error evaluation strategy, where a unidirectional edge is
evaluated twice so that the control flow is similar between different
edges.

However, this does not seem to materially improve runtime for cases when
most edges are indeed bidirectional; for models with a high number of
unidirectional edges this has a visible negative impact on performance.