gltfpack: Introduce artificial seam edges on UV flips
When two triangles with reverse UV direction meet, an edge between them
needs to be a seam edge. This makes sure that UV does not degrade by
preventing the collapses that go across different UV directions.

Ideally this should be implemented inside meshopt library at some point,
but it requires a separate function and some way to make possibly-manifold
vertices "seam" which may run into issues with the seam handling logic.

So for now, pre-process the mesh by splitting vertices in gltfpack, and
remap triangles with negative UV area to use the newly created vertices.
The number of these vertices is generally low.