gltfpack: Remove split vertices after simplification
After simplification concludes, the simplified mesh is still using the
split vertices, which results in a higher number of vertices overall as
we do not reindex the mesh after simplification. We do optimize it, which
includes removal of unused vertices, so it's sufficient to change the index
buffer after simplification to refer to the original copy of the split.