tools: Add -test mode to codectest
In this mode, we take a set of file paths, load each one and recompress
it. The file names should have the form:

	vN_sK_name_etc

... where N is the number of vertices and K is the vertex stride. Each
file may either be compressed using vertex codec or uncompressed. The
file will be decompressed, recompressed, and various stats will be
measured.

Right now this will return meaningless (zero) stats as the compression
algorithm is the same; this, however, will be useful to develop the next
version of vertex codec.