vertexcodec: Introduce an experimental version 0xe
This version is *not* for production; we allow to set it via the
official API to be able to test and incrementally develop it, but the
format will not be finalized for a while. When it is, the version number
will be changed to 1, and existing calls for 0xe / data with 0xe will
fail to encode/decode.