vertexcodec: Generalize encodeBytes to support bit group palettes
Instead of hardcoding bit counts, we can now encode using arbitrary 4
bit counts. When using the original v0 values, this produces the exact
same output, but this gives us an opportunity to select a different set
of bits inside the same block.

As part of this, bits=1 is changed to bits=0; bits=0 was always correct,
as bits=1 refers to a 1-bit mask plus sentinels.