vertexcodec: Initial implementation of byte group controls
In v0 format, we encode using 0, 2, 4 and 8-bit groups. This provides a
reasonable coverage, but misses some other useful bit groups; notably,
1-bit groups (with sentinel encoding) can be used to more efficiently
encode mostly-0 sequences without spending an extra 2 bytes on bit
masks, and 6-bit groups (with sentinel encoding) provide a middle ground
between 4-bit and 8-bit that can be useful for data that is otherwise
difficult to compress.

While we could choose one of 6 groups for every 16 bytes instead of one
of 4, this noticeably increases the header overhead. Instead, we can
preselect a few 4-element arrangements for bits, and use control bits to
select the right one.

For now we choose a hard-coded palette of 3 entries. The 4-th entry may
be useful for literal encoding to save space on header bits for very
incompressible blocks. The palette itself is subject to change.

To select the palette entry, we allocate 2-bit control metadata for
every byte channel in the vertex block. We will likely need more bits in
the future to accomodate deltas et al, but for now let's start with
this.

For now we pick the palette by trying all three variants and choosing
the shorter option. That results in measuring every byte group 12 times;
it should be possible to do this faster if we measure every byte group 6
times and aggregate the results.