vertexcodec: Update kBitsV1 table for slightly better compression
This table was found by bruteforcing all 455 permutations of 15 possible
bit combinations in each entry; the best result was ~4.90% reduction,
but for now we pick the version with 0/2/4/8 as one of the entries
(which is ~4.83% reduction) as that makes it easier to retain
compatibility with v0. The previous table that this code replaces got
~4.59% reduction.

The new table also makes more logical sense: 0/1/2/4 optimizes for
well-compressible data, and 1/4/6/8 optimizes for poorly-compressible
data. 2/4/6/8 would perhaps make more sense, but 1/4/6/8 performs
better.