vertexcodec: Implement parsing of control blocks for scalar decode
This is very straightforward: just need to skip the control bits for v0,
and use them to select the bit table. To make decoding a little simpler,
for now we mandate that kBitsV1[0] == kBitsV0; this makes it possible to
index the bit table unconditionally.

Note that this doesn't include the byte group decoding, so we can't
fully parse experimental encoding yet.