vertexcodec: Support 1-bit and 6-bit groups for scalar decoder
Instead of reading one byte at a time we now read one or three bytes at
a time. This results in a decoding structure that is still fairly
generic but able to decode 1-bit and 6-bit groups with the same macros.

The preliminary performance analysis suggests that clang generates
roughly as efficient code as it used to, and gcc is a little behind.
Wasm code size increases by ~500 bytes post-gzip (~9%?). These all will
need to be reevaluated after we implement SIMD decoding anyhow - for
example, decodeBytesGroup could be implemented with more loops to
optimize for code size instead if we target platforms with fallbacks
like Wasm.