vertexcodec: Implement initial support for expanded encoding with SSE
This change introduces decodeBytesGroupSimdX which can decode all bit
widths we need in the new encoding using SSE. Code for 0/2/4/8 bits is
copied from decodeBytesGroupSimd, and code for 1/6 bits is new.

To decode 1-bit groups, we can mostly just use the existing tables.
Unfortunately, the bits in these tables are reversed compared to what we
need; this does not affect counts, but this does affect shuffle masks.

For now we perform a byte reverse using scalar math for this.
Alternative options exist, including reworking encoding to flip the
bits, or alternative table construction.

For 6-bit groups, we mostly just need to move the 6-bit values into the
right place after a byte shuffle. We try to take advantage of the bit
arrangement to do this with fewer ands; better implementation schemes
may exist.

It's not clear yet how to do a latency optimization codepath for either
of these groups in a way that is faster than not doing it. This is
perhaps less important for 1-bit groups as the masks are directly
present in source data; 6-bit groups present challenges wrt value
realignment.

Additionally, the dispatch structure for switch here runs into more
optimization concerns; to minimize these for now we split the v1
decoding completely into a separate function with the same structure.
Ideally we should be able to improve performance of "X" variant so that
we can merge these back.