vertexcodec: Implement NEON support for decodeBytesSimdX
The code closely mirrors the SSE decoding strategy; for 6-bit groups,
we need a 16-byte table lookup -- if possible, we use vqtbl1q_u8, which
requires AArch64; the fallback involves splitting the loads and table
lookups in two.

For 1-bit groups, we do bit reverse using the same scalar sequence as on
Intel; we should have access to bit reverse intrinsic which may make
decoding faster, but since we might reverse the encoded data we can wait
to use it.