vertexcodec: Implement Wasm SIMD version of new decoding
This is a straightforward port of SSE version with no surprises.
Blissfully this appears to run almost as fast as the original version,
at least on node/v8, so this might be a case where it's easy to just
share the code with the existing decoder. Even without this the size
expansion is tolerable.