vertexcodec: Initial implementation of AVX512 extended decoding
Thanks to the power of multishift, we can decode 1-bit groups using the
same code and an adjusted table entry. For 6-bit groups, we run into a
problem where we were assuming 8 bytes are enough to hold all 16 values,
but for 6-bit groups we need 12 bytes. So we now read a full 16 bytes
and shuffle it using an extra per-bit shuffle mask.