vertexcodec: Tag default cases in SIMD switches as unreachable
To avoid the compiler inserting extra bounds checks for bits dispatch,
we now tag default: case as unreachable in SIMD implementations.

This is safe because the values are going to be bounded statically, and
all offsetting we will do will keep the value in bounds handled by
switch.

This was less necessary before, because the compiler knew that the
dispatch value (bitslog2) was 2-bit so all cases were handled; it's
going to be less clear with the new dispatch structure.

We could do this for the scalar implementation in the future but there's
not a strong need for that.