vertexcodec: Implement initial support for v1 SIMD dispatch
We do not have the implementation for decodeBytesGroupSimd yet, but the
outer framing can be parsed the same way we do this for scalar code.

Note that it is safe to call memset/memcpy on aligned vertex count
(buffer has enough space to write to, and data has enough space due to
tail). For now though use unaligned counts for simplicity as this should
only affect the last block.