vertexcodec: Integrate v1 decoding for SSE implementation
Due to the extreme requirements for the decoding loop we need to somehow
integrate the dispatch with minimal changes and minimal extra
computations for the switch index.

For now, map the entire bit group range linearly such that each index
for v0 maps to 0..3 as before, and v1 maps to 4..7 for one palette entry
and 5..8 for another.

On gcc this approach results in extra bounds checks around the switch,
which for some reason is not eliminated via SIMD_UNREACHABLE macro; on
clang this results in reasonable lowering.