vertexcodec: Adjust decoding for performance on small streams
kByteGroupDecodeLimit was bumped to 28 to accommodate 6-bit groups, but
since they were removed we can go back to 24 to get slightly tighter
loops on small inputs.

For the last block, vertex count is often unaligned; if it is literal or
zero, we can overcopy or overset because everything is properly aligned
and input data must have at least 16b of tail anyway.