Implement delta modes (without auto tuning)
We now process deltas 4 at a time to make modes easier to implement, and
allow per-4b mode for diffs. If 4b deltas end up not being useful it
might be possible to select 2+1+1 configuration but for now that seems
like overkill, as we expect that in 2b mode consecutive 2 bytes in 4
byte group encode a component of a similar structure.