Implement delta tuning
This bruteforces delta modes per 4b, keeping the smallest output; we
redundantly reencode the data which should not matter for tuning but
it's easier to implement that way.

Since we operate externally to the actual encoding, we have to pick the
delta mode per channel for the entire data stream, not per block. It's
unclear if this is good or bad, but might actually be good as the
modality should depend on the component semantics. This also improves
zstd gains by favoring consistent choices.