gltfpack: Output buffers[] list after extension list
Semantically these are exactly the same - JSON output is unordered.
However, when inspecting the files manually, it makes more sense to
see assets/extensions first. We could append buffers at the end, but we
try to follow logical order wrt referencing glTF objects earlier in the
JSON stream just in case streaming parsers want to take the opportunity
to optimize the parsing flow.