gltfpack: Support writing animation tracks with custom time
This allows more flexibility in processing the animations before
serialization. When the track time is present, we write it instead of a
fixed-rate frequency track.

For now no attempt to deduplicate time tracks is taken; this could be
added in the future if need be, but the assumption is that the more
optimal way to serialize tracks with shared time ranges is to use
resampling anyway.