vertexcodec: Increase tail size for experimental encoding
When new encoding is selected, in addition to storing the first vertex
in the tail we also store channel data. Channel data will encode the
transformation type for every 4 bytes of the vertex; for now we only
need to encode a 2-bit value, but we reserve a full byte for simplicity
- and we may need 2 bytes for this in the future to accomodate rotates.

While varying the channel data per vertex block may be helpful, it
increase the block overhead and switching encodings for adjacent blocks
penalizes efficiency of LZ compressors. As such, we will only work with
a single encoding across the entire data stream for every channel.