vertexcodec: Implement channel estimation
When using the new version, we can now pick the optimal channel
encoding; for now, we do this in a fairly naive fashion: we encode
deltas in all possible ways and count the number of bytes.

This does not, strictly speaking, minimize the output size: for
simplicity, we estimate the number of bytes using a specific fixed bit
group layout, which seems pretty close in terms of its efficiency for
any delta encoding type. This is still not very fast, but hopefully
better, faster heuristics can be developed in the future.