vertexcodec: Implement SSE support for new delta decoding
In addition to accumulating the deltas via 16-bit or 32-bit adds (or
xors), we need to perform zigzag decoding in larger units. This is more
complicated, because for 8-bit data we do this before the transpose to
fully utilize SIMD units; thus, 16 and 32-bit unzigzag needs to work on
decomposed integer values.

For 16-bit, it's easy enough to unpack and repack these; for 32-bit,
repacking is more involved due to the lack of unsigned 32-bit packing
instructions. So we opt for doing a byte-wise unzigzag instead, which
just needs to propagate the bits when shifting the value by 1 across
byte boundaries manually.