vertexcodec: Switch to byte-wise unzigzag16 for SSE
For unclear reasons, despite seemingly better codegen, unzigzag16 with
unpacks is a little faster on clang compared to the byte-wise version.
However, it's noticeably slower on MSVC in a way that affects top line
decode bandwidth, so let's opt for using a better version to avoid
surprises.

Also switch to using _mm_andnot_si128 explicitly; some compilers can
automatically transform this but it's easy to do ourselves for safety.