vertexcodec: Remove support for 32-bit deltas
Both xors and 32-bit deltas are not particularly profitable on real
data; of the two, 32-bit deltas take more effort to support whereas xor
is not only easier to support but also decodes faster.

In the future we might keep xor mode to be able to select faster deltas
for channels that do not compress at all anyway.