Implement delta refs encoding sketch
We estimate the best vertex to encode each vertex in relation to; the
lookback distance is limited to 4 bits, and can't cross block boundaries
to simplify decoding. The refs are stashed before all other data since
they are necessary to decode; we could split them by block in the future
to make streaming easier I guess.