vertexcodec: Reduce the valid rotation range to 0..7
The entire 32-bit rotation space is unnecessary for our purpose:
rotating by 8 yields the same compression as rotating by 0, as it just
changes the byte order and bytes are compressed individually.

This reduces the channel encoding consumption to 5 bits (2+3) and
reduces the iteration during encoding 4x. We could technically encode
using 4 bits (using 0, 1, 2..9 values...), but this doesn't meaningfully
impact tail sizes so maybe isn't that necessary.