vertexcodec: Use faster rotation estimation
Instead of bruteforcing all rotations and computing the xor-encoded
size, we can estimate the size without using encoding, and do this
analysis on groups of 16 deltas. For each group, we OR the deltas
together, which results in "1" bits whenever there's any inconsistency
within the group.

This is an estimate that is not always as good as the correct value, but
it's generally pretty close: if there is a significant bias towards
selecting a specific rotation, this algorithm will still find it, and do
so 10x faster.