vertexcodec: Optimize estimateControl by reusing measurements
Instead of calling a generic encodeBytesMeasure twice, we can note that
it redundantly recomputes almost all of the information: in v1, bits
1/2/4 are available to both control modes, so we just need to compute 0
and 8 sizes - of which, 8 is a constant size. So we can compute the
sizes for both streams in parallel.

Doing this preserves the bitstream exactly, but results in ~20% faster
encoding at level 1.