vertexcodec: Use branchless append in encodeBytesGroup
The sentinel branch is difficult to predict; since we have enough space
to encode all bytes as sentinels due to decode limit padding, it is safe
to append every byte unconditionally and move the pointer for out of
range values. This accelerates all levels of encoding further, up to 30%
for levels 0 and 1.

Also rework encodeBytes flow to call the function just once; this mostly
just makes sure the compiler can inline it without issues, as otherwise
the function is too large to be inlined into two separate paths.