vertexcodec: Inline encodeBytesMeasure into estimateChannel
After previous changes, encodeBytesMeasure is no longer used by any
other function than estimateChannel. Inlining the function into
estimateChannel allows us to simplify the code, and improves
optimizations as an explicit measure is faster vs table selection in
practice. This also allows us to drop one of the bit group modes to gain
extra performance in the future.

In addition we also fix last_vertex handling (this was incorrectly using
the first vertex for all blocks instead of last vertex of the previous
block) and reduce memset overhead by limiting it to the last (partial)
block.