vertexcodec: Accelerate channel estimation via block skipping
Instead of analyzing every block we could look at a subset of blocks and
assume that the statistics of data in different blocks is reasonably
close. This is a little brute-force, but gets almost the same
compression results on a variety of files, so for now we can do this
unconditionally at every level, which significantly increases the
encoding throughput of levels 2 and 3.