demo: Add special tests for v1 format
While v1 is not finalized yet, hopefully the only changes that will
happen here are going to be to the vertex number, and/or channel
encoding.

To make sure we can decode the data in all cases this adds two tests,
one with a normal v1 stream encoded at level 3, which does not exercise
all format features, and a specially crafted stream that exercises all
bitgroup widths, all control modes and all channel modes to encode the
same data.