vertexcodec: Reduce tail padding for v1
When v0 was implemented, the decode limit was thought to be 32 bytes;
hence the tail had to be at least 32 bytes to cover the decode overhang.
Since then, decode limit was revised to 24 bytes (8 bytes of 4-bit data
plus 16 bytes of outliers), but the tail remained the same as the format
is fixed. For v1, the decode limit is the same and this is unlikely to
change for future revisions either since 6-bit groups did not end up
working out (they would have required 28 bytes of tail space), so it is
time to reduce the tail to 24 bytes for v1.

This only impacts small blobs but makes the format a little cleaner.