vertexcodec: Change new version from 0xe to 1
With all the previous changes, the encoding format is now more or less
stable. While it is possible to expand the encoding a little more to
enhance XOR encoding in some cases, these are unlikely and should not
invalidate any existing v1 streams.

As part of this, -ce setting in gltfpack is enabled in release builds
for easier testing.