vertexcodec: Implement a new parallel delta decoding scheme
Instead of using serialized delta processing, we can decode deltas
in groups of 4 in parallel, using available processor resources better
by removing the serial dependencies. This computes a prefix sum that
we then have to complete by doing a few serial additions.

For now this is only implemented for NEON; on Apple M2, this results
in ~20% end to end speedup for decoding, suggesting the deltas are
specifically faster. In future commits the rest of the code will
hopefully be reworked to match so that we can clean up the older code.

An alternative scheme is to do full 16-byte prefix sums, which also
requires fixup at the end, but also is harder to generalize to 16-bit
and 32-bit undelta. Fortunately that approach is also a little slower.