vertexcodec: Remove redundant splat for first group
The contents of lanes 1-3 do not matter for any tN, so we can simply
copy the vector which saves an instruction. NEON was already applying
this optimization, but SSE/Wasm weren't.

Also change Wasm store to use store_lane; the compiler already
synthesized this instruction, but it's better to be explicit.