vertexcodec: Implement alternative delta decoding acceleration
While computing prefix sums partially in parallel works on Apple Mn, it is
unfortunately slower on weaker CPUs like Graviton2. The equivalent approach
is also slower on Zen3, so this would require maintaining a special version
just for Apple.

The reason why parallel prefix sums work better is that they allow for more
parallelism during the loop iteration, at the expense of a little more work.
Alternatively we can try to overlap individual loop iterations - there is
a data dependency on "pi" which serializes the long latency chain between
iterations. What we can do instead is recompute pi for the next iteration
based on rN, which is easier than a full prefix sum, and results in even
larger performance improvement on Apple M2.