vertexcodec: Enable NEON latency optimization on Windows
On Qualcomm Snapdragon X CPUs, this optimization helps significantly
just as it does for Apple CPUs, boosting decoding from ~3.72 GB/s to
4.54 GB/s. To keep the scope of this optimization to cases where we know
it does help, enable this for Windows-on-ARM builds as well.

This optimization is known to slightly regress throughput on Graviton 2
as well as Cortex-A78, so for now we still keep it disabled for Android
and Linux builds.