tools: Add cumulative size stats to codectest
As v1 is getting wrapped up, v0-v1 comparison is less valuable compared
to the absolute sizes. It might make sense to replace the cumulative
ratio improvement stats with something like a ratio geomean, but for now
add total size (which is skewed towards larger input files, but is an
accurate representation of total disk size or load time impact for a
given set).