demo: Rework Nanite demo to link to Metis dynamically
Having separate build configurations makes it harder to work with this
prototype, as different configurations may require rebuilds. Long term
the plan is to remove the Metis dependency, but working through that
requires new algorithms with a stable baseline, so we need this code for
now.

To make things easier, this change removes dependency on metis.h and
uses dynamic linking instead; we only need three functions and a couple
defines so this is fairly painless.