demo: Use a mostly-non-recursive triangle clustering
When Metis is used for triangle clustering, we use recursive
partitioning to try to split the mesh into two groups each of which is
split further. This is tricky to balance, as the earlier splits are
constraining future splits, and if the groups aren't evenly divisible by
cluster size, they are bound to get fewer and fewer triangles.

What we can do instead is to use non-recursive partitioning; it will
sometimes leave larger clusters so we can target a slightly larger
partition count to reduce the chance of that happening, and split
overly large clusters (that are rare) further recursively.

This interacts in a strange way with retries, so for now keep using
recursive clusterization.