clusterizer: Implement sparse mode for buildTriangleAdjacency
When clusterizer is used to generate meshlets for a small subset of the
mesh, the bulk of the work is in O(vertex_count) loops inside adjacency
build as opposed to the actual clusterizer. Normally the mesh has more
indices than vertices, but for mesh subsets this relationship may be
reversed.

This problem is solved in simplifier by allocating a sparse remap; while
that solution is optimal, it is harder to implement here as the index
buffer needs to be copied and remapped, and this also requires hashing
and other helpers.

A seemingly simple solution is to compute the index range used by the
subset and only allocate and compute adjacency for it; unfortunately, it
is common to have the index range span most of the vertex buffer as the
indices are not necessarily local.

For now this change implements a middleground: we still allocate
adjacency for the entire vertex buffer, but we only fill a subset used
by the index buffer. This requires tracking which vertices we filled,
which we do via a spare bit on the counts[] array that can be set in the
first pass and cleared in the second.

When using 64 triangle clusters on a 1M triangle mesh, this makes the
clusterization itself ~2x faster, and the hierarchical clustering
process can be ~25% faster as a result.