clusterizer: Separate sparse and dense adjacency builds
In sparse mode we still had a memset(counts) that cleared the entire
array; this is not necessary in sparse mode and can be meaningfully
slower especially if the region of memory is fresh and the pages need to
be committed.

We can instead clear every used index in a separate pass. This results
in a small further performance improvement. This also results in sparse
builds using almost completely separate code, at which point it's
cleaner to split it into a separate function, which is what this commit
also does.