demo: Add connected component count statistics
For hierarchical clusterization, disconnected clusters that happen early
in the pipeline are problematic because unless they get merged with
other clusters to fill the gap, they will create an increasing number of
locked edges downstream. This makes this a good comparison metric
between different DAG algorithms, so we now compute this for every level
separately.

Unlike main demo, for now we compute this based on position-only
connectivity as this is what is used for cluster partitioning as well.