demo: Remove recursive Metis triangle clusterizer
With the added metrics it is easy to compare recursive to non-recursive
clusterization; recursive clusterization results in more accumulated
cluster splits over time. It yields slightly more triangles per cluster
on average, but from the global perspective the results are similar.

In the interest of keeping just one superior implementation we will keep
the non-recursive variant, although it's not fully clear which one of
these is better overall.