demo: Add ideal DAG depth metric
This helps, at a glance, to see if the DAG is truncated early or is too
deep compared to the "optimal" depth. The optimal depth by itself is not
necessarily a good target to chase, as long as the real depth is within
a couple levels the DAG should be just fine.