partition: Expand merge scoring heuristics
Instead of just the shared counts, we can now use the external count
heuristic when picking the merge target, as well as a group size based
heuristic that prioritizes smaller groups to produce more even
partitions.

To make this a little easier, countExternal can now be called on two
isolated groups before merging.