partition: Use cluster offsets to access cluster indices
Instead of building a vector of Cluster structures, replace this with
direct indexing and use a prebuilt offset table that we build from
counts.

This bakes an assumption that the cluster indices are always going to be
sequential in memory for the input, but this should be easy to adjust if
the external interface were to change.