partition: Rewrite cluster adjacency build using arrays
Instead of using nested STL data structures, this change rewrites
cluster adjacency build using arrays with offsets/counts.

This is significantly more involved than basic triangle adjacency: we
need to build two data structures, a temporary index->cluster map and a
permanent cluster adjacency structure that also records shared vertex
count.

For the former, we use a simple adjacency structure, with a caveat that
it's important to only record each vertex in each cluster once to reduce
memory use (which would otherwise be total_index_count, but computing a
precise count allows ~4-5x reduction on average).

For the latter, things are complicated: not only can we not estimate the
adjacency list sizes precisely without computing them, the computation
requires making a union of all cluster lists for every vertex which is
both tricky and slow.

For now we instead use a moderately conservative estimate: for each
cluster, assume the lists in each vertex do not overlap so we will need
to allocate space for all of them. This usually ends up overallocating
~4x space for adjacency but since this is efficiently packed and avoids
the need to compute the same lists twice, it seems like a reasonable
compromise for now.

Throughout this process we use used[] to temporarily flag the duplicate
vertices in each cluster, which requires resetting the flag back; we
might need used[] in the future in other parts of the code, otherwise
there might be more efficient ways to implement this.

The result is ~10x faster than the STL build (using glibc allocator) for
the adjacency build, and reduces the end-to-end hierarchical build by
~25% when meshopt clusterizer is used.