partition: Replace map with heap for tracking group order
Instead of storing groups as a map of vectors, we use vector of vectors
(which will be replaced in future commits) and track the order in a
separate heap. While we in theory need to remove the group we merge into
the smallest group from the heap, in practice we can defer this removal
- the group will be empty so we can wait until we encounter it in heap
order.

This is a little faster compared to using map, and removes the final
dependency on the most complex data structure used in the prototype.