partition: Replace nested vectors for group storage with lists
Instead of storing every group contiguously, use a list embedded in an
array. This makes it easier to merge groups and can also be used to
store per-cluster data, which eliminates part[] array as well.

With this, the last remnants of the STL structures are gone; the code is
not fully final because we need to pick the actual metrics used for
scoring and ordering, and some of the existing code will need to be
reworked once it's fully done as well.