vertex/indexcodec: Add meshopt_decodeVertex/IndexVersion
In some cases it may be useful to check the encoded blob's version ahead
of time, for example to make sure it's encoded using the latest version,
or to make sure that vertex codec v0 is used for glTF data.

Note that these functions just check the headers; they may return a valid
version number even if the input is malformed.

While we're at it, unify the maximum version checks for both codecs.