partition: Reduce partition size range
With our existing heuristics, we would on average produce partitions
with size ~15% over the target. This is not entirely desirable and skews
the comparison metrics somewhat; it's especially problematic when the
target partition size is a power of two, since it creates an easy path
to get to the max partition size by incrementally doubling until the
target, and then merging partitions with target and target/2 together.

For now just reducing the extra size allowance a little to 3/8 of target
maintains reasonable partitioning properties with closer-to-target
average size.