partition: Adjust group scoring to normalize shared counts
Instead of applying a group size bias, which is somewhat arbitrary as it
conflates the dimensions (of cluster counts and vertex counts), we now
normalize the shared count by inverse square root of group sizes. Since
we, ideally, expect the boundary of a cluster group to have sqrt(V)
vertices, this results in a normalized account of shared vertex count;
while it is a little redundant to use rsqrt for the group we're merging,
this keeps the scoring mostly symmetric.

We can also use the product of rsqrt's which matches the metric used in
KaPPa ["Engineering a Scalable High Quality Graph Partitioner" (2010)].
It doesn't seem to strongly matter one way or the other, and the weights
are a little easier to debug and understand with + so we'll use that for now.