demo: Extract all metrics output into a separate function
In addition to significantly cleaning up the main processing function,
this also ensures that the metrics are consistent, being gathered from
the same level. Before this we actually had an issue where some metrics
were gathered after simplification and some - before; as a result, some
values such as boundary will change between levels of detail compared to
the previous output.