demo: Update simplify demo with support for grid view
When multiple files are being opened, we now show and simplify them all
side by side, using shared settings and shared camera. This makes it
easier to iterate on simplification algorithm as the impact on multiple
scenes can be visible concurrently.

As a byproduct this also moves simplification stats to HTML labels in
each view and tweaks some other parts of the code accordingly.