clusterizer: Add buildMeshletsFlex with flexible cluster sizing
In some cases, producing clusters from two disconnected regions of the
mesh may result in efficiency loss in the downstream algorithms; this is
sometimes hard to avoid due to unpredictable traversal flow. This change
introduces a variant of the existing meshlet builder,
meshopt_buildMeshletsFlex, which has two triangle count limits: a lower
and upper limit. When a cluster is below a lower limit, the logic
remains the same, however for larger clusters we allow splitting the
resulting cluster if not doing that would result in excessively large
volume, as measured by split_factor.

For now, split_factor is applied to the expected radius to serve as a
cutoff value; this assumes somewhat uniform input triangle sizes, but
also protects against clusters that are already disjointly merged (due
to splits while they are under the minimum limit) becoming even larger.

Additionally, it is now possible to switch to a sizing heuristic that
favors axis aligned cluster bounds; to do that, a negative cone_weight
must be passed, since in these cases cone weighting is not useful. This
is especially valuable for ray tracing workloads when an efficient BVH
must be built over the resulting clusters.